Bench: 1835286

+++ src/search.h
@@ -17,7 +17,7 @@ extern int LMP_MARGIN[MAX_PLY][2];
 
 void initReductions();
 
-uint64_t perft(Board& board, int depth);
+uint64_t perft(Board& board, int16_t depth);
 
 struct SearchParameters {
     bool perft; // Perft (requires depth)
@@ -32,7 +32,7 @@ struct SearchParameters {
     uint64_t winc; // White's increment per move (ms)
     uint64_t binc; // Black's increment per move (ms)
     int movestogo; // Moves to the next time control
-    int depth; // Search depth
+    int16_t depth; // Search depth
     uint64_t nodes; // Search exactly this many nodes
     int mate; // TODO: Search for mate in X moves
     uint64_t movetime; // Search exactly this many ms
